code/pyproject.toml
[tool.poetry]
name = "pricing"
version = "0.1.0"
description = ""
authors = ["Peter Cliff <peter.cliff@jisc.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.7"
jirafs-matplotlib = "^1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
